move byte array extensions into their own files, retarget to netcoreapp3.1 and netstandard2.1
